backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
slowapi>=0.1.8

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
redis>=4.5.4
alembic>=1.10.2

# Document processing
pymupdf>=1.22.3
python-docx>=0.8.11

# Vector storage
weaviate-client>=3.15.4

# Recommendation system
lightfm>=1.17
numpy>=1.24.2
scipy>=1.10.1
scikit-learn>=1.2.2

# Testing
pytest>=7.3.1
httpx>=0.24.0

# Utils
python-dotenv>=1.0.0
tenacity>=8.2.2 